use server-side rendered chart in dashboard widget too
