+++ src/class-dashboard-widget.php
@@ -25,26 +25,6 @@ class Dashboard_Widget
         add_meta_box('koko-analytics-dashboard-widget', 'Koko Analytics', array($this, 'dashboard_widget'), 'dashboard', 'side', 'high');
     }
 
-    public function get_script_data(): array
-    {
-        $stats = new Stats();
-        $dateStart = create_local_datetime('-14 days')->format('Y-m-d');
-        $dateEnd = create_local_datetime('now')->format('Y-m-d');
-        return array(
-            'root' => rest_url(),
-            'nonce' => wp_create_nonce('wp_rest'),
-            'i18n' => array(
-                'Visitors' => __('Visitors', 'koko-analytics'),
-                'Pageviews' => __('Pageviews', 'koko-analytics'),
-            ),
-            'startDate' => $dateStart,
-            'endDate' => $dateEnd,
-            'data' => array(
-                'chart' => $stats->get_stats($dateStart, $dateEnd, 'day'),
-            ),
-        );
-    }
-
     public function dashboard_widget(): void
     {
         $number_of_top_items = (int) apply_filters('koko_analytics_dashboard_widget_number_of_top_items', 5);
@@ -52,6 +32,10 @@ class Dashboard_Widget
         $dateToday = create_local_datetime('today, midnight')->format('Y-m-d');
         $realtime = get_realtime_pageview_count('-1 hour');
         $totals = $stats->get_totals($dateToday, $dateToday);
+
+        $dateStart = create_local_datetime('-14 days');
+        $dateEnd = create_local_datetime('now');
+        $chart_data = $stats->get_stats($dateStart->format('Y-m-d'), $dateEnd->format('Y-m-d'), 'day');
 
         if ($number_of_top_items > 0) {
             $posts = $stats->get_posts($dateToday, $dateToday, 0, $number_of_top_items);
