+++ src/views/dashboard-widget.php
@@ -3,13 +3,15 @@
 /**
 * @var int $number_of_top_items
 * @var int $realtime
+* @var array $chart_data
 * @var array $posts
 * @var array $referrers
 * @var stdClass $totals
-* @var $dateStart*
-*
+* @var \DateTimeInterface $dateStart
+* @var \DateTimeInterface $dateEnd
 */
 
+use KokoAnalytics\Chart_View;
 ?>
 <link rel="stylesheet" href="<?php echo plugins_url('assets/dist/css/dashboard.css', KOKO_ANALYTICS_PLUGIN_FILE); ?>?v=<?php echo KOKO_ANALYTICS_VERSION; ?>">
 <script src="<?php echo plugins_url('assets/dist/js/dashboard-widget.js', KOKO_ANALYTICS_PLUGIN_FILE); ?>?v=<?php echo KOKO_ANALYTICS_VERSION; ?>" defer></script>
@@ -21,13 +23,12 @@
     </p>
 </div>
 
-<div id="ka-dashboard-widget-chart" style="display: none; margin-top: 2em;">
+<div id="ka-dashboard-widget-chart" style="margin-top: 2em;">
     <h3>
        <?php echo esc_html__('Showing site visits over last 14 days', 'koko-analytics'); ?>
     </h3>
-    <div id="koko-analytics-dashboard-widget-mount">
-        Please wait, your chart is loading. <br />
-        If nothing shows up, check your browser console for any error messages.
+    <div id="ka-chart-2">
+        <?php new Chart_View($chart_data, $dateStart, $dateEnd); ?>
     </div>
 </div>
 
@@ -72,7 +73,3 @@
         <?php echo esc_html__('View all statistics', 'koko-analytics'); ?>
     </a>
 </p>
-
-<script>
-var koko_analytics = <?php echo json_encode($this->get_script_data()); ?>;
-</script>
